Moved integrationtest method to the DsImageClient integration test
