+++ src/test/java/dk/kb/image/util/DsImageClientTest.java
@@ -14,10 +14,18 @@
  */
 package dk.kb.image.util;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.Mockito.mockStatic;
 
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
+
+import javax.imageio.ImageIO;
+import javax.servlet.http.HttpServletResponse;
+import javax.ws.rs.core.StreamingOutput;
 
 import org.apache.cxf.jaxrs.utils.JAXRSUtils;
 import org.apache.cxf.message.MessageImpl;
@@ -25,10 +33,13 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import dk.kb.image.config.ConfigAdjuster;
 import dk.kb.image.config.ServiceConfig;
+import dk.kb.util.Resolver;
 import dk.kb.util.oauth2.KeycloakUtil;
 import dk.kb.util.webservice.OAuthConstants;
 
@@ -38,10 +49,10 @@ import dk.kb.util.webservice.OAuthConstants;
 @Tag("integration")
 public class DsImageClientTest {
     private static final Logger log = LoggerFactory.getLogger(DsImageClientTest.class);
-   
+
     private static DsImageClient remote = null;
     private static String dsImageDevel=null;  
-    
+
     @BeforeAll
     static void setUp() throws Exception{
         try {
@@ -53,20 +64,20 @@ public class DsImageClientTest {
             log.error("Integration yaml 'ds-image-integration-test.yaml' file most be present. Call 'kb init'"); 
             fail();
         }
-        
+
         try {            
             String keyCloakRealmUrl= ServiceConfig.getConfig().getString("integration.devel.keycloak.realmUrl");            
             String clientId=ServiceConfig.getConfig().getString("integration.devel.keycloak.clientId");
             String clientSecret=ServiceConfig.getConfig().getString("integration.devel.keycloak.clientSecret");                
             String token=KeycloakUtil.getKeycloakAccessToken(keyCloakRealmUrl, clientId, clientSecret);           
             log.info("Retrieved keycloak access token:"+token);            
-            
+
             //Mock that we have a JaxRS session with an Oauth token as seen from within a service call.
             MessageImpl message = new MessageImpl();                            
             message.put(OAuthConstants.ACCESS_TOKEN_STRING,token);            
             MockedStatic<JAXRSUtils> mocked = mockStatic(JAXRSUtils.class);           
             mocked.when(JAXRSUtils::getCurrentMessage).thenReturn(message);
-                                                                         
+
         }
         catch(Exception e) {
             log.warn("Could not retrieve keycloak access token. Service will be called without Bearer access token");            
@@ -76,7 +87,45 @@ public class DsImageClientTest {
 
     @Test
     public void test() throws IOException {
-    //Must be one unit test to test the setup method is working 
+        //Must be one unit test to test the setup method is working 
     }
-        
+
+
+
+    @Test
+    public void testPlaceholderImageStreamingOutput() throws IOException {
+        //This is not a DsImageClient test,  but logic will call ds-license service.          
+        HttpServletResponse response = Mockito.mock(HttpServletResponse.class);
+        if (Resolver.getPathFromClasspath("ds-image-integration-test.yaml") == null){
+            fail("Internal test config is not present. 'testPlaceholderImageStreamingOutput' is therefore not run. Please update aegis and do 'kb init' to make this run.");
+        }
+
+        try (ConfigAdjuster configAdjuster = new ConfigAdjuster("ds-image-integration-test.yaml")){
+            // Get test image
+            ByteArrayOutputStream testImageByteArray = getImageAsByteArrayOS("nonExisting.jpg");
+            // Get image through ImageAccessValidation
+            StreamingOutput streamingImage = ImageAccessValidation.handleNoAccessOrNoImage("notExistingImage", response, false);
+
+            // Convert image to ByteArrayOutputStream for comparison
+            ByteArrayOutputStream streamedImageByteArray = new ByteArrayOutputStream();
+            streamingImage.write(streamedImageByteArray);
+
+            assertEquals(testImageByteArray.size(), streamedImageByteArray.size());
+        }
+    }
+
+
+
+    private static ByteArrayOutputStream getImageAsByteArrayOS(String imgName) throws IOException {
+        String imgPath = Resolver.getPathFromClasspath(imgName).toString();
+
+        BufferedImage trueImage = ImageIO.read(new File(imgPath));
+        ByteArrayOutputStream trueImageByteArray = new ByteArrayOutputStream();
+
+        // Write the BufferedImage to the ByteArrayOutputStream in the specified format
+        ImageIO.write(trueImage, "jpg", trueImageByteArray);
+        trueImageByteArray.flush();
+        return trueImageByteArray;
+    }
+
 }
